implement remaining check methods
