+++ server/src/service/business/checkService.ts
@@ -1,6 +1,7 @@
 import { IChecksRepository, IMonitorsRepository } from "@/repositories/index.js";
 import type { MonitorType, MonitorStatusResponse, CheckErrorInfo, Check } from "@/types/index.js";
 import type { HardwareStatusPayload, PageSpeedStatusPayload } from "@/types/network.js";
+import { AppError } from "@/utils/AppError.js";
 import { ParseBoolean } from "@/utils/utils.js";
 
 const SERVICE_NAME = "checkService";
@@ -8,25 +9,21 @@ const SERVICE_NAME = "checkService";
 class CheckService {
 	static SERVICE_NAME = SERVICE_NAME;
 
-	private db: any;
 	private errorService: any;
 	private monitorsRepository: IMonitorsRepository;
 	private checksRepository: IChecksRepository;
 	private logger: any;
 	constructor({
-		db,
 		errorService,
 		monitorsRepository,
 		logger,
 		checksRepository,
 	}: {
-		db: any;
 		errorService: any;
 		monitorsRepository: IMonitorsRepository;
 		logger: any;
 		checksRepository: IChecksRepository;
 	}) {
-		this.db = db;
 		this.errorService = errorService;
 		this.monitorsRepository = monitorsRepository;
 		this.logger = logger;
@@ -132,15 +129,6 @@ class CheckService {
 
 		const result = await this.checksRepository.findByMonitorId(monitorId, sortOrder, dateRange, filter, parsedPage, parsedRowsPerPage, parsedStatus);
 
-		// const result = await this.db.checkModule.getChecksByMonitor({
-		// 	monitorId,
-		// 	sortOrder,
-		// 	dateRange,
-		// 	filter,
-		// 	page,
-		// 	rowsPerPage,
-		// 	status,
-		// });
 		return result;
 	};
 
@@ -151,14 +139,10 @@ class CheckService {
 			throw this.errorService.createBadRequestError("No team ID in request");
 		}
 
-		const checkData = await this.db.checkModule.getChecksByTeam({
-			sortOrder,
-			dateRange,
-			filter,
-			page,
-			rowsPerPage,
-			teamId,
-		});
+		const parsedPage = page ? parseInt(page) : page;
+		const parsedRowsPerPage = rowsPerPage ? parseInt(rowsPerPage) : rowsPerPage;
+
+		const checkData = await this.checksRepository.findByTeamId(sortOrder, dateRange, filter, parsedPage, parsedRowsPerPage, teamId);
 		return checkData;
 	};
 
@@ -166,8 +150,7 @@ class CheckService {
 		if (!teamId) {
 			throw this.errorService.createBadRequestError("No team ID in request");
 		}
-
-		const summary = await this.db.checkModule.getChecksSummaryByTeamId({ teamId });
+		const summary = await this.checksRepository.findSummaryByTeamId(teamId);
 		return summary;
 	};
 
@@ -183,7 +166,7 @@ class CheckService {
 		// For verificaiton, throws an error if monitor doesn't belong to team
 		await this.monitorsRepository.findById(monitorId, teamId);
 
-		const deletedCount = await this.db.checkModule.deleteChecks(monitorId);
+		const deletedCount = await this.checksRepository.deleteByMonitorId(monitorId);
 		return deletedCount;
 	};
 	deleteChecksByTeamId = async ({ teamId }: { teamId: string }) => {
@@ -191,14 +174,12 @@ class CheckService {
 			throw this.errorService.createBadRequestError("No team ID in request");
 		}
 
-		const deletedCount = await this.db.checkModule.deleteChecksByTeamId(teamId);
+		const deletedCount = await this.checksRepository.deleteByTeamId(teamId);
 		return deletedCount;
 	};
 
 	updateChecksTTL = async ({ teamId, ttl }: { teamId: string; ttl: string }) => {
-		const SECONDS_PER_DAY = 86400;
-		const newTTL = parseInt(ttl, 10) * SECONDS_PER_DAY;
-		await this.db.checkModule.updateChecksTTL(teamId, newTTL);
+		throw new AppError({ message: "Not implemented", service: SERVICE_NAME, method: "updateChecksTTL", status: 500 });
 	};
 }
